pyproject.toml
[tool.poetry]
name = "news-parser"
version = "0.1.0"
description = ""
authors = ["Sherif Ragimov <ragimov700@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pillow = "^10.1.0"
lxml = "^4.9.4"
flake8 = "^6.1.0"
setuptools = "^69.0.3"
python-telegram-bot = "13.7"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
